Refactoring Check and Radio
* Not fully decoupled because Radio is using some of Check code.
* Shoes3/shoes3#333
