+++ shoes/native/gtk.c
@@ -1013,7 +1013,7 @@ VALUE shoes_native_control_get_tooltip(SHOES_CONTROL_REF ref) {
     return rb_str_new2(gtk_widget_get_tooltip_text(GTK_WIDGET(ref)));
 }
 
-static gboolean shoes_button_gtk_clicked(GtkButton *button, gpointer data) {
+gboolean shoes_button_gtk_clicked(GtkButton *button, gpointer data) {
     VALUE self = (VALUE)data;
     shoes_control_send(self, s_click);
     return TRUE;
@@ -1115,57 +1115,6 @@ double shoes_native_slider_get_fraction(SHOES_CONTROL_REF ref) {
 
 void shoes_native_slider_set_fraction(SHOES_CONTROL_REF ref, double perc) {
     gtk_range_set_value(GTK_RANGE(ref), perc);
-}
-
-SHOES_CONTROL_REF shoes_native_check(VALUE self, shoes_canvas *canvas, shoes_place *place, VALUE attr, char *msg) {
-    SHOES_CONTROL_REF ref = gtk_check_button_new();
-    // set visual state before connecting signal
-    if (RTEST(ATTR(attr, checked))) {
-        gtk_toggle_button_set_active(GTK_TOGGLE_BUTTON(ref), TRUE);
-    }
-
-    if (!NIL_P(shoes_hash_get(attr, rb_intern("tooltip")))) {
-        gtk_widget_set_tooltip_text(GTK_WIDGET(ref), RSTRING_PTR(shoes_hash_get(attr, rb_intern("tooltip"))));
-    }
-
-    g_signal_connect(G_OBJECT(ref), "clicked",
-                     G_CALLBACK(shoes_button_gtk_clicked),
-                     (gpointer)self);
-    return ref;
-}
-
-VALUE shoes_native_check_get(SHOES_CONTROL_REF ref) {
-    return gtk_toggle_button_get_active(GTK_TOGGLE_BUTTON(ref)) ? Qtrue : Qfalse;
-}
-
-void shoes_native_check_set(SHOES_CONTROL_REF ref, int on) {
-    // bug264 - don't toggle if already set to desired state.
-    gboolean new_state;
-    new_state = on ? TRUE : FALSE;
-    if (gtk_toggle_button_get_active(GTK_TOGGLE_BUTTON(ref)) != new_state)
-        gtk_toggle_button_set_active(GTK_TOGGLE_BUTTON(ref), new_state);
-}
-
-SHOES_CONTROL_REF shoes_native_radio(VALUE self, shoes_canvas *canvas, shoes_place *place, VALUE attr, VALUE group) {
-    SHOES_CONTROL_REF ref;
-    GSList *list = NULL;
-    if (!NIL_P(group)) {
-        shoes_control *lctrl;
-        VALUE leader = rb_ary_entry(group, 0);
-        Data_Get_Struct(leader, shoes_control, lctrl);
-        list = gtk_radio_button_get_group(GTK_RADIO_BUTTON(lctrl->ref));
-    }
-
-    ref = gtk_radio_button_new(list);
-
-    if (!NIL_P(shoes_hash_get(attr, rb_intern("tooltip")))) {
-        gtk_widget_set_tooltip_text(GTK_WIDGET(ref), RSTRING_PTR(shoes_hash_get(attr, rb_intern("tooltip"))));
-    }
-
-    g_signal_connect(G_OBJECT(ref), "clicked",
-                     G_CALLBACK(shoes_button_gtk_clicked),
-                     (gpointer)self);
-    return ref;
 }
 
 VALUE shoes_native_clipboard_get(shoes_app *app) {
