+++ shoes/native/gtk/gtkswitch.h
@@ -4,7 +4,7 @@
 #define SHOES_GTK_SWITCH_H
 
 // There is some magical way to create this but let's do this manually for now
-SYMBOL_ID(active);
+SYMBOL_EXTERN(active);
 
 SHOES_CONTROL_REF shoes_native_switch(VALUE self, shoes_canvas *canvas, shoes_place *place, VALUE attr, char *msg);
 void shoes_native_switch_set_active(SHOES_CONTROL_REF ref, gboolean activate);
